Fixing changes requested by @louis-ver
